add breakpoints & span size selection
creates breakpoints based on max-width screen size. Upon hitting the breakpoint
the selected column will take up the full width of it's row, and any
other elements on that row will be moved down to the next row. Also adds the ability to select the amount of columns per row each element
should take up, and adds this same feature in conjunction with responsive breakpoints.